requirements.txt
mcp>=0.9.0
# HTTP client and ASGI testing transport
httpx==0.27.2
requests>=2.31.0

# Vector database client
qdrant-client>=1.7.0

# FastAPI stack (versions compatible together)
fastapi==0.110.0
pydantic>=2.12,<3.0
pydantic-settings>=2.2,<3.0
anyio>=4.4,<5.0
python-multipart>=0.0.9
uvicorn>=0.29,<0.30

# ML stack (unchanged)
transformers>=4.35.0
sentence-transformers>=2.2.2
torch>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
ollama>=0.1.0

# Embedded LLM (Phi-4 Mini via llama.cpp)
# NOTE: llama-cpp-python is installed separately in Dockerfile with CUDA support
# Uncomment below only for CPU-only local development:
# llama-cpp-python>=0.2.90

# Wave 6 - Real embeddings support
nltk>=3.8.0
tqdm>=4.65.0

# Dev / Quality / Governance
ruff==0.5.5
mypy==1.11.2
pytest==9.0.1
pytest-asyncio>=1.3.0
pre-commit==3.7.1
cyclonedx-python-lib==4.2.2
pip-licenses==4.3.3

# Audio I/O Layer (PRD Section 8.12)
# STT: Vosk (Kaldi-based, offline, CPU-only, 94-95% WER)
vosk>=0.3.45

# TTS: Coqui TTS (Glow-TTS + HiFi-GAN, MOS 4.2-4.4)
TTS>=0.22.0

# Audio processing utilities
soundfile>=0.12.1
librosa>=0.10.0

# Monitoring and Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-starlette>=0.48b0
starlette[cors]>=0.27.0
opentelemetry-instrumentation-httpx>=0.48b0
grpcio>=1.60.0

# Database (PostgreSQL with async support)
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
# ─────────────────────────────────────────────────────────────────────────────
# System Monitoring and WebSocket Support (Task 4)
# ─────────────────────────────────────────────────────────────────────────────
psutil>=5.9.0                          # System metrics (CPU, RAM, disk, network)
GPUtil>=1.4.0                          # GPU monitoring (optional)
websockets>=12.0                       # WebSocket protocol support
python-socketio>=5.10.0                # Socket.IO for real-time communication
python-engineio>=4.8.0                 # Engine.IO transport layer